requirements.txt
ansible==2.9.6
ansible-runner==1.4.7
appnope==0.1.0
asgiref==3.2.10
backcall==0.1.0
bcrypt==3.2.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==4.4.0
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.11.0
docutils==0.17.1
idna==2.9
importlib-metadata==2.0.0
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.13.3
Jinja2==2.11.2
lockfile==0.12.2
lxml==4.5.1
Markdown==3.3.2
MarkupSafe==1.1.1
paramiko==2.7.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psutil==5.8.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.1
PyMySQL==1.0.2
PyNaCl==1.4.0
python-daemon==2.3.0
pytz==2020.1
PyYAML==5.4.1
requests==2.23.0
six==1.12.0
sqlparse==0.4.1
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.7
zipp==3.3.1
